bot/requirements.txt
python-telegram-bot==21.6
httpx==0.27.0
uvloop==0.19.0
pytz